# Let library users handle output however they like, and add flatgeobuf output