--- a/Cargo.toml
+++ b/Cargo.toml
@@ -13,10 +13,12 @@ anyhow = "1.0.52"
 clap = { version = "3.0.0", features = ["derive"] }
 # TODO Separate library/binary dependencies
 csv = "1.1.6"
+flatgeobuf = { version = "3.26.1", default-features = false }
 fs-err = "2.6.0"
 geo = "0.18.0"
 geo-types = "0.7.2"
 geojson = { version = "0.22.2", features = ["geo-types"] }
+geozero = { version = "0.10.0", default-features = false, features = ["with-geojson"] }
 ordered-float = "3.4.0"
 rand = "0.8.4"
 rstar = "0.8.4"