--- a/Cargo.toml
+++ b/Cargo.toml
@@ -9,17 +9,17 @@ license = "Apache-2.0"
 readme = "README.md"
 
 [dependencies]
-anyhow = "1.0.52"
+anyhow = "1.0.72"
 clap = { version = "3.0.0", features = ["derive"] }
 # TODO Separate library/binary dependencies
-csv = "1.1.6"
+csv = "1.2.2"
 flatgeobuf = { version = "3.26.1", default-features = false }
-fs-err = "2.6.0"
-geo = "0.18.0"
-geo-types = "0.7.2"
-geojson = { version = "0.22.2", features = ["geo-types"] }
+fs-err = "2.9.0"
+geo = "0.26.0"
+geo-types = "0.7.11"
+geojson = { version = "0.24.1", features = ["geo-types"] }
 geozero = { version = "0.10.0", default-features = false, features = ["with-geojson"] }
-ordered-float = "3.4.0"
+ordered-float = "3.7.0"
 rand = "0.8.4"
-rstar = "0.8.4"
-serde_json = "1.0.73"
+rstar = "0.11.0"
+serde_json = "1.0.104"