3 functional changes here:

1) Change the public API to output features through a callback, letting the caller do whatever they like. Closes #50.
2) Add an option to output the result as flatgeobuf instead of geojson
3) Huge speedups -- from 58s to 22s for one large benchmark -- just from upgrading to latest geo packages!

I used the input data from #5: `time cargo run --release -- disaggregate --od-csv-path ~/Downloads/od_sp_center.csv --zones-path ~/Downloads/zones_sp_center.geojson --output-path result.fgb --output-fgb`

Before this PR, output to geojson took 58s. After this PR, down to 22s!

FGB output takes much longer, 50s. Not sure why yet, though I have a few ideas. The resulting file is much smaller -- 735MB FGB, 1.4GB GeoJSON. I haven't checked downstream yet to see which is faster to read, or if reading is even the bottleneck in a larger pipeline.

This is a step towards https://github.com/dabreegster/odjitter/issues/49#issuecomment-1645620084 improving the API for language bindings, but the input is still a bunch of files.